backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
google-genai>=0.5.0
arxiv>=2.1.0
requests>=2.31.0
jinja2>=3.1.0
